% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_clinvar.R
\name{download_clinvar}
\alias{download_clinvar}
\title{Import ClinVar VCF}
\usage{
download_clinvar()
}
\value{
Latest version of Clinvar VCF
}
\description{
This function allows you to read a ClinVar VCF into a table
}
