% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P3STAT294.R
\docType{package}
\name{P3STAT294}
\alias{P3STAT294}
\title{P3STAT294: A package for downloading ClinVar dataset}
\description{
The \code{P3STAT294} package provides one important function, namely, \code{download_clinvar()}.
}
\section{\code{P3STAT294} functions}{

The \code{download_clinvar()} function downloads the ClinVar dataset from
'ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar.vcf.gz'
}

